ai_service/requirements.txt
# AI Image Generation Service - Python Dependencies

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Gemini API
google-generativeai==0.3.1

# Stable Diffusion and ML
diffusers==0.24.0
transformers==4.35.2
torch>=2.0.0
accelerate==0.25.0
safetensors>=0.4.0

# Image processing
Pillow==10.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Optional: For better performance
# xformers==0.0.23  # Uncomment if using CUDA GPU
